optimize u64 to u32

+++ arrow-array/src/array/byte_view_array.rs
@@ -552,8 +552,8 @@ impl<T: ByteViewType + ?Sized> GenericByteViewArray<T> {
 
             // The lower 32 bits encode the length (little-endian),
             // the upper 96 bits hold the actual data
-            let l_len = l_bits as u32 as u64;
-            let r_len = r_bits as u32 as u64;
+            let l_len = l_bits as u32;
+            let r_len = r_bits as u32;
 
             // Remove the length bits, leaving only the data
             let l_data = l_bits >> 32;
